Implemented the physics engine's broad phase and narrow phase collision detection systems.
The physics engine now responds to collisions with a two step process, detection(broad phase) and resolution(narrow phase). T

+++ engine/source/gom/source/src/Game_object_manager.hpp
@@ -25,7 +25,7 @@ namespace gom {
 
 	class  Game_object_manager final {
 	public:
-		typedef uint16_t						type_id;
+		typedef uint32_t						type_id;
 		typedef std::map<type_id, Creator*>		creator_map;
 		typedef std::vector<Game_object*>		vpgame_objects;
 		typedef std::vector<Game_object_handle> vgame_object_handles;
